# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Thoth API",
    "version": "0.1.0",
    "description": "Live OSINT layers, intelligence, watches and on-demand lookups. See docs/reference/api.md for conventions (authentication, list and error shapes, the event object).",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4000",
      "description": "API, direct (host only)"
    },
    {
      "url": "/",
      "description": "Through the app (access token when gated)"
    }
  ],
...
  "paths": {
...
    "/api/stream": {
      "get": {
        "operationId": "getStream",
        "tags": [
          "System"
        ],
        "summary": "Live Server-Sent Events",
        "description": "Events: connected, snapshot, layer_changed, heartbeat. Resume with `known` (base64 JSON {layer: version}) or Last-Event-ID.",
        "parameters": [
          {
            "name": "known",
            "in": "query",
            "required": false,
            "description": "base64 JSON {layer: version} to resume from",
            "schema": {
              "type": "string",
              "maxLength": 20000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
    "/api/layers/{layer}/export": {
      "get": {
        "operationId": "getLayersLayerExport",
        "tags": [
          "Layers"
        ],
        "summary": "Download the layer (CSV or GeoJSON)",
        "parameters": [
          {
            "name": "layer",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "csv",
                "geojson"
              ],
              "default": "csv"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/brief": {
      "get": {
        "operationId": "getBrief",
        "tags": [
          "Intelligence"
        ],
        "summary": "Deterministic brief: counts and top items",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
    "/api/theaters": {
      "get": {
        "operationId": "getTheaters",
        "tags": [
          "Intelligence"
        ],
        "summary": "Theatre presets (region, cities, feeds)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/search": {
      "get": {
        "operationId": "getSearch",
        "tags": [
          "Intelligence"
        ],
        "summary": "Full-text search over events",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search text",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          {
            "name": "layer",
            "in": "query",
            "required": false,
            "description": "Restrict to one layer",
            "schema": {
              "type": "string",
              "maxLength": 40
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/imagery": {
      "get": {
        "operationId": "getImagery",
        "tags": [
          "Intelligence"
        ],
        "summary": "Freshest low-cloud Sentinel-2 scene over a point",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "description": "Latitude",
            "schema": {
              "type": "number",
              "minimum": -90,
              "maximum": 90
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "description": "Longitude",
            "schema": {
              "type": "number",
              "minimum": -180,
              "maximum": 180
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        }
      }
    },
    "/api/analytics/trend": {
      "get": {
        "operationId": "getAnalyticsTrend",
        "tags": [
          "Intelligence"
        ],
        "summary": "Daily series with the sitrep archive",
        "parameters": [
          {
            "name": "layer",
            "in": "query",
            "required": false,
            "description": "Layer id",
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 90,
              "default": 14
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/watch": {
      "get": {
        "operationId": "getWatch",
        "tags": [
          "Watches and notifications"
        ],
        "summary": "List watches",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      },
      "post": {
        "operationId": "postWatch",
        "tags": [
          "Watches and notifications"
        ],
        "summary": "Create a watch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "kind": {
                        "type": "string",
                        "enum": [
                          "keyword",
                          "layer",
                          "severity"
                        ]
                      },
                      "value": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 200
                      },
                      "note": {
                        "type": "string",
                        "maxLength": 500,
                        "default": ""
                      }
                    },
                    "required": [
                      "kind",
                      "value"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "kind": {
                        "const": "area"
                      },
                      "value": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 200
                      },
                      "note": {
                        "type": "string",
                        "maxLength": 500,
                        "default": ""
                      },
                      "lat": {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90
                      },
                      "lon": {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180
                      },
                      "radius_km": {
                        "type": "number",
                        "exclusiveMinimum": 0,
                        "maximum": 2000
                      },
                      "geom": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "enum": [
                              "Polygon",
                              "MultiPolygon"
                            ]
                          },
                          "coordinates": {
                            "type": "array",
                            "items": {},
                            "minItems": 1
                          }
                        },
                        "required": [
                          "type",
                          "coordinates"
                        ]
                      }
                    },
                    "required": [
                      "kind",
                      "value"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "security": [
          {
            "bearerWriteKey": []
          },
          {
            "headerWriteKey": []
          }
        ]
      }
    },
    "/api/watch/{id}": {
      "delete": {
        "operationId": "deleteWatchId",
        "tags": [
          "Watches and notifications"
        ],
        "summary": "Delete a watch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "security": [
          {
            "bearerWriteKey": []
          },
          {
            "headerWriteKey": []
          }
        ]
      }
    },
    "/api/watch/matches": {
      "get": {
        "operationId": "getWatchMatches",
        "tags": [
          "Watches and notifications"
        ],
        "summary": "Recent events matching any watch",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/notify": {
      "post": {
        "operationId": "postNotify",
        "tags": [
          "Watches and notifications"
        ],
        "summary": "Send a Telegram message",
        "description": "Reports `disabled` without a bot token and chat id.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 3500,
                    "description": "Message text"
                  }
                },
                "required": [
                  "text"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "security": [
          {
            "bearerWriteKey": []
          },
          {
            "headerWriteKey": []
          }
        ]
      }
    },
    "/api/sitrep": {
      "post": {
        "operationId": "postSitrep",
        "tags": [
          "Sitreps"
        ],
        "summary": "Build and archive today's sitrep",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "security": [
          {
            "bearerWriteKey": []
          },
          {
            "headerWriteKey": []
          }
        ]
      },
      "get": {
        "operationId": "getSitrep",
        "tags": [
          "Sitreps"
        ],
        "summary": "Latest sitrep",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
      "BadRequest": {
        "description": "Invalid parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Missing or invalid write key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limited: retry after `Retry-After` seconds",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          }
        }
      },
      "Upstream": {
        "description": "The upstream behind this lookup failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerWriteKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API_WRITE_KEY, for POST and DELETE"
      },
      "headerWriteKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Thoth-Key"
      }
    }
  }
}